ERT-699: Removed call to ensemble_clear_internalization()
As part of the function enkf_main_init_run() the function
ensemble_config_init_internalization() was called to clear the
internalization flag of all the nodes. This had the uninted consequence
of clearing the internalization flag of the gen_data nodes, thereby
loosing the REPORT_STEPS: config which has been loaded from the
configuration file.