Silence warnings for usage of printf without format
GNU compiler puts an attribute on the printf family of statements to
statically check the format string to see if it matches up with the
arguments which are passed.

If just a char* is passed and no further arguments, you'll get a
"warning: format not a string literal and no format arguments" since
the string may contain formatting specifiers (%something) but there
is nothing on the stack.

These invocations should be changed to having an "%s" format string
so that the argument is not further parsed.